requirements.txt
fastapi
uvicorn[standard]
python-multipart
paddleocr>=3.3.0
paddlepaddle==3.2.2
opencv-python-headless
Pillow
opencc-python-reimplemented
